requirements.txt
langchain_google_genai
google-generativeai
python-dotenv
pytesseract
torchvision
streamlit
pyttsx3
torch